+++ examples/undo/main.rs
@@ -1,15 +1,19 @@
 use crate::rules::*;
 use std::io::Read;
+use weasel::ability::ActivateAbility;
 use weasel::battle::Battle;
 use weasel::creature::{CreateCreature, CreatureId};
+use weasel::entity::EntityId;
 use weasel::event::EventTrigger;
+use weasel::round::{EndRound, StartRound};
 use weasel::team::{CreateTeam, TeamId};
 use weasel::Server;
 
 mod rules;
 
-static TEAM_ID: TeamId<CustomRules> = 0;
-static CREATURE_ID: CreatureId<CustomRules> = 0;
+const TEAM_ID: TeamId<CustomRules> = 0;
+const CREATURE_ID: CreatureId<CustomRules> = 0;
+const ENTITY_ID: EntityId<CustomRules> = EntityId::Creature(CREATURE_ID);
 
 fn main() {
     print_intro();
@@ -26,6 +30,10 @@ fn print_intro() {
     println!("Example to demonstrate how to undo/redo player actions with weasel.");
     println!("Move around the soldier on the battlefield.");
     println!();
+    print_controls();
+}
+
+fn print_controls() {
     println!("  Controls:");
     println!("    w - Move up");
     println!("    s - Move down");
@@ -33,12 +41,13 @@ fn print_intro() {
     println!("    a - Move left");
     println!("    u - Undo");
     println!("    r - Redo");
+    println!("    h - Display the controls");
     println!("    q - Quit");
 }
 
 fn game_loop() {
     // Create a server.
-    let server = create_server();
+    let mut server = create_server();
     println!();
     display_world(&server);
     // Main loop.
@@ -53,23 +62,30 @@ fn game_loop() {
         if let Some(key) = input {
             match key {
                 'w' => {
+                    walk(&mut server, Direction::Up);
                     display_world(&server);
                 }
                 's' => {
+                    walk(&mut server, Direction::Down);
                     display_world(&server);
                 }
                 'd' => {
+                    walk(&mut server, Direction::Right);
                     display_world(&server);
                 }
                 'a' => {
+                    walk(&mut server, Direction::Left);
                     display_world(&server);
                 }
                 'u' => {
+                    server = undo(server);
                     display_world(&server);
                 }
                 'r' => {
+                    server = redo(server);
                     display_world(&server);
                 }
+                'h' => print_controls(),
                 'q' => break,
                 _ => {}
             }
@@ -92,5 +108,31 @@ fn create_server() -> Server<CustomRules> {
     CreateCreature::trigger(&mut server, CREATURE_ID, TEAM_ID, Square { x: 0, y: 0 })
         .fire()
         .unwrap();
+    server
+}
+
+/// Moves the creature on step towards the given direction.
+fn walk(server: &mut Server<CustomRules>, direction: Direction) {
+    // Start a round.
+    StartRound::trigger(server, ENTITY_ID).fire().unwrap();
+    // Activate the 'walk' ability of the creature.
+    let result = ActivateAbility::trigger(server, ENTITY_ID, WALK)
+        .activation(direction)
+        .fire();
+    // We print an error in case the movement is not allowed.
+    if result.is_err() {
+        println!("Movement not allowed!");
+    }
+    // End the round.
+    EndRound::trigger(server).fire().unwrap();
+}
+
+/// Undo the last action.
+fn undo(server: Server<CustomRules>) -> Server<CustomRules> {
+    server
+}
+
+/// Redo the last undoed action.
+fn redo(server: Server<CustomRules>) -> Server<CustomRules> {
     server
 }