+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Example for undo/redo of events.
+- The methods `activable`, `on_round_start` and `on_round_end` now take `BattleState` as argument.
 
 ## [0.4.1] - 2020-02-22
 ### Changed